\name{parallelDGP-package}
\alias{parallelDGP-package}
\alias{parallelDGP}
\docType{package}
\title{\packageTitle{parallelDGP}}
\description{\packageDescription{parallelDGP}}
\section{Package Content}{\packageIndices{parallelDGP}}
\author{\packageAuthor{parallelDGP}}
\section{Maintainer}{\packageMaintainer{parallelDGP}}
\keyword{package}
